Move sql_execute callback to Explorer.Backend.DataFrame
- Add callback to Explorer.Backend.DataFrame behaviour
- Implement in Explorer.PolarsBackend.DataFrame and LazyFrame
- Remove callback from Explorer.Backend
- Update sql/2 to pick backend from first DataFrame or use default for empty tables
