+++ lib/explorer/backend/data_frame.ex
@@ -258,6 +258,13 @@ defmodule Explorer.Backend.DataFrame do
   @callback concat_columns([df], out_df :: df()) :: df
   @callback concat_rows([df], out_df :: df()) :: df
 
+  # SQL
+
+  @callback sql_execute(
+              tables :: [{binary(), df}],
+              sql_string :: binary()
+            ) :: df
+
   # Groups
 
   @callback summarise_with(df, out_df :: df(), aggregations :: [{column_name(), lazy_series()}]) ::
