+++ lib/explorer/polars_backend/data_frame.ex
@@ -959,6 +959,25 @@ defmodule Explorer.PolarsBackend.DataFrame do
     %{out_df | data: out_data}
   end
 
+  # SQL
+
+  @impl true
+  def sql_execute(tables, sql_string) do
+    tables_with_df =
+      Enum.map(tables, fn {name, df} ->
+        {name, df.data}
+      end)
+
+    alias Explorer.PolarsBackend.{Native, Shared}
+
+    with {:ok, polars_ldf} <- Native.sql_execute(tables_with_df, sql_string),
+         {:ok, polars_df} <- Native.lf_compute(polars_ldf) do
+      Shared.create_dataframe!(polars_df)
+    else
+      {:error, error} -> raise error
+    end
+  end
+
   # Groups
 
   @impl true
