+++ lib/explorer/polars_backend/lazy_frame.ex
@@ -644,6 +644,26 @@ defmodule Explorer.PolarsBackend.LazyFrame do
     %{out_df | data: polars_df}
   end
 
+  # SQL
+
+  @impl true
+  def sql_execute(tables, sql_string) do
+    tables_with_df =
+      Enum.map(tables, fn {name, df} ->
+        {name, df.data}
+      end)
+
+    alias Explorer.PolarsBackend.Native
+
+    with {:ok, polars_ldf} <- Native.sql_execute(tables_with_df, sql_string),
+         {:ok, names} <- Native.lf_names(polars_ldf),
+         {:ok, dtypes} <- Native.lf_dtypes(polars_ldf) do
+      Explorer.Backend.DataFrame.new(polars_ldf, names, dtypes)
+    else
+      {:error, error} -> raise error
+    end
+  end
+
   @impl true
   def concat_columns([%DF{data: head_data} | tail], %DF{} = out_df) do
     out_ldf_data = Shared.apply(:lf_concat_columns, [[head_data | Enum.map(tail, & &1.data)]])
